Consolidate and fix tests
